+++ Kernel/Arch/HAL.h
@@ -5,7 +5,7 @@
 // Defines the Hardware Abstraction Layer (HAL),
 // and implements it depending on the requested architecture ARCH_*
 
-#if !defined(ARCH_x86) && !defined(ARCH_arm64)
+#if !defined(x86_64) && !defined(arm64)
 #error "Invalid architecture provided ! Please define a valid architecture maccro from the supported list."
 #endif
