+++ Kernel/Arch/x86_64/IDT.asm
@@ -1,0 +1,17 @@
+bits 64
+
+; void setIDT(IDT_LocationDescriptor_64* descriptor);
+global setIDT
+setIDT:
+	push rbp
+	mov rbp, rsp
+
+	; load IDT ; address given by 'descriptor' arg
+	; mov eax, [ebp + 8]
+	; lidt [eax]
+	lidt [rdi]
+
+	mov rsp, rbp
+	pop rbp
+	ret
+; END setIDT
